src/requirements.txt
discord.py==2.3.2
python-dotenv==0.19.2
py-cord==2.4.1
requests==2.31.0
